# packages/backend/docs/vcwSwagger.yml
openapi: 3.0.3
info:
  title: Vcw4Next
  description: |-
    🚧🚧
  version: "0.0"
servers:
  - url: "localhost:8081/v1/"
  - url: "http://vcw-app.develop.s4eo.com/"
...
paths:
...
  /projects/{project_id}/vcws/{vcw_id}/diagnostics:
    get:
      tags:
        - Diagnostics page (swot fields)
      summary: Shows all diagnostics of the given vcw. 🟢
      description: TODO
      operationId: getByVcw
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/examples/diagnostic-new'
        '404':
          description: Diagnostic not found
          content:
            'text/plain; charset=utf-8':
              schema:
                type: string
                example: 'Diagnostic not found'
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: diagnostic_id
          in: path
          description: "diagnostic id"
          required: true
          schema:
            type: integer
    post:
      tags:
        - Diagnostics page (swot fields)
      summary: Creates a single diagnostic. 🟢
      description: TODO
      operationId: save
      responses:
        '201':
          description: Created diagnostics
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/examples/diagnostics-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..."}
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: swotField
          in: formData
          description: "swot field: strength, weakness, opportunities, threat"
          required: true
          schema:
            type: string
        - name: name
          in: formData
          description: 'name'
          required: true
          schema:
            type: string
        - name: description
          in: formData
          description: "description"
          required: true
          schema:
            type: string
...
  /projects/{project_id}/vcws/{vcw_id}/ideas/:
    get:
      tags:
        - Ideas page
      summary: Gets all ideas of the vcw. 🟢
      description: TODO
      operationId: getIdeas
      responses:
        '201':
          description: Created ideas
          content:
            application/json:
              schema:
                type: array
                item:
                $ref: '#/components/examples/ideas-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..."}
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
    post:
      tags:
        - Ideas page
      summary: Adds new idea 🟢
      description: TODO
      operationId: save
      responses:
        '201':
          description: Created idea
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/examples/ideas-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: entryTypeId
          in: formData
          description: "Type of entry: user (1), vcw_project (2) or third_party (3)."
          required: true
          schema:
            type: Long
            example: 3
        - name: source
          in: formData
          description: idea source
          required: true
          schema:
            type: object
            example:
              source:
                zomato: string
                url: string
  /projects/{project_id}/vcws/{vcw_id}/ideas/{idea_id}:
    put:
      tags:
        - Ideas page
      summary: Edits given idea 🟢
      description: TODO
      operationId: update
      responses:
        '201':
          description: Updated idea
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/examples/ideas-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: idea_id
          in: path
          description: "idea id"
          required: true
          schema:
            type: integer
        - name: entryTypeId
          in: formData
          description: "Type of entry: user (1), vcw_project (2) or third_party (3)."
          required: true
          schema:
            type: Long
            example: 3
        - name: source
          in: formData
          description: idea source
          required: true
          schema:
            type: object
            example:
              source:
                zomato: string
                url: string
    delete:
      tags:
        - Ideas page
      summary: Deletes given idea 🟢
      description: TODO
      operationId: delete
      responses:
        '400':
          description: Invalid idea id supplied
        '404':
          description: Idea not found
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
        - name: idea_id
          in: path
          description: Idea id to delete
          required: true
          schema:
            type: integer
            format: int64
  /projects/{project_id}/vcws/{vcw_id}/criterias/:
    get:
      tags:
        - Criterias page
      summary: Gets all criterias of the vcw. 🚧
      description: TODO
      operationId: getCriterias
      responses:
        '201':
          description: Created criterias
          content:
            application/json:
              schema:
                type: array
                $ref: '#/components/examples/criteria-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
    post:
      tags:
        - Criterias page
      summary: Adds new criteria 🚧
      description: TODO
      operationId: save
      responses:
        '201':
          description: Created criteria
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/examples/criteria-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: entryTypeId
          in: formData
          description: "Type of entry: user (1), vcw_project (2) or third_party (3)."
          required: true
          schema:
            type: Long
            example: 3
        - name: valueType
          in: formData
          description: "Type of value, it can be 'number' or 'yes_or_no'"
          required: true
          schema:
            type: string
            example: number
        - name: source
          in: formData
          description: criteria source
          required: true
          schema:
            type: object
            example:
              source:
                zomato: string
                url: string
  /projects/{project_id}/vcws/{vcw_id}/criterias/{criteria_id}:
    put:
      tags:
        - Criterias page
      summary: Updates given criteria 🚧
      description: TODO
      operationId: update
      responses:
        '201':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/examples/criteria-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: criteria_id
          in: path
          description: "criteria id"
          required: true
          schema:
            type: integer
        - name: entryTypeId
          in: formData
          description: "Type of entry: user (1), vcw_project (2) or third_party (3)."
          required: true
          schema:
            type: Long
            example: 3
        - name: valueType
          in: formData
          description: "Type of value, it can be 'number' or 'yes_or_no'"
          required: true
          schema:
            type: string
            example: number
        - name: source
          in: formData
          description: criteria source
          required: true
          schema:
            type: object
            example:
              source:
                zomato: string
                url: string
    delete:
      tags:
        - Criterias page
      summary: Deletes given criteria 🚧
      description: TODO
      operationId: delete
      responses:
        '400':
          description: Invalid criteria id supplied
        '404':
          description: Criteria not found
      parameters:
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
        - name: criteria_id
          in: path
          description: Criteria id to delete
          required: true
          schema:
            type: integer
            format: int64
  /projects/{project_id}/vcws/{vcw_id}/ideasAndCriterias:
    get:
      tags:
        - Purification and population page
      summary: Gets all ideaAndCriteria pairs available for the current vcw 🚧
      description: TODO
      operationId: getByVcw
      responses:
        '201':
          description: Created criterias
          content:
            application/json:
              schema:
                type: array
                $ref: '#/components/examples/ideas-and-criterias-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: project_id
          in: path
          description: "project id"
          required: true
          schema:
            type: integer
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
    post:
      tags:
        - Purification and population page
      summary: Adds new ideaAndCriteria pair to vcw 🚧
      description: TODO
      operationId: save
      responses:
        '201':
          description: Created ideaAndCriteria pair-
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/examples/ideas-and-criterias-new'
        '400':
          description: Invalid arguments
          content:
            application/json:
              example: {
                "timestamp": "2023-01-31T12:03:11.417+00:00",
                "status": 400,
                "error": "Bad Request",
                "trace": "...MethodArgumentNotValidException: Validation failed for..." }
      parameters:
        - name: project_id
          in: path
          description: "project id"
          required: true
          schema:
            type: integer
        - name: vcw_id
          in: path
          description: "vcw id"
          required: true
          schema:
            type: integer
        - name: ideaId
          in: formData
          description: "Selected idea id"
          required: true
          schema:
            type: Long
            example: 3
        - name: criteriaId
          in: formData
          description: "Selected criteria id"
          required: true
          schema:
            type: Long
            example: 3
        - name: value
          in: formData
          description: "Associated value"
          required: true
          schema:
            type: Float
            example: 3
        - name: source
          in: formData
          description: Idea and criteria source
          required: false
          schema:
            type: object
            example:
              source:
                zomato: string
                url: string
...
components:
...
  examples:
...
    ideas-new:
      properties:
        name:
          type: string
          example: 'Italian food is tasty'
        entryTypeId:
          type: Long
          example: 3
        source:
          type: object
          $ref: '#/components/examples/source-new'
      xml:
        name: 'ideas-new'
    criteria-new:
      properties:
        name:
          type: string
          example: Price
        entryTypeId:
          type: Long
          example: 3
        valueType:
          type: string
          example: number
        source:
          type: object
          $ref: '#/components/examples/source-new'
      xml:
        name: 'criteria-new'
    source-new:
      properties:
        name:
          type: string
          example: zomato
        description:
          type: string
          example: null
        url:
          type: string
          example: www.zomato.com
      xml:
        name: 'source-new'
    diagnostics-new:
      properties:
        swotField:
          type: string
          example: threats
        name:
          type: string
          example: 'crabs'
        description:
          type: string
          example: 'They are chasing me right now, help!'
      xml:
        name: 'diagnostics-new'